tests/integration: extract shared UnreliableIo crash-simulation module
Pull the crash-simulating IO out of checkpoint_crash_atomicity.rs into a
shared tests/integration/unreliable_io.rs module so other tests can
simulate crashes and storage failures without duplicating the machinery.
This is the test infrastructure half of PR #8103, extracted so the
upcoming power-loss recovery test (issue #7995) lands on a clean
foundation.

The module models unreliable storage in general, not one failure kind:
it tracks the gap between issued writes and fsync-covered bytes, and a
test picks the failure to materialize — process crash (nothing lost),
power loss (only durable bytes survive), or power loss mid-fsync with
torn writeback. Injected I/O errors and reordered writeback belong here
too as tests grow to need them.

The API is generalized while moving: arm_crash_on_db_sync becomes
arm_crash_on_sync and the snapshot counters drop their db_ prefix, since
any file can be the crash target, and durable_disk_snapshot becomes
durable_files. durable_files returns per-path byte images so a test can
materialize the post-crash disk state and reopen it with the regular
platform IO, which is exactly what the power-loss test needs.

Tests: cargo test -p core_tester --test integration_tests checkpoint_crash;
cargo clippy -p core_tester --all-features --all-targets
