+++ tests/integration/mod.rs
@@ -25,6 +25,7 @@ mod stmt_journal;
 mod stmt_readonly;
 mod storage;
 mod trigger;
+mod unreliable_io;
 mod wal;
 
 #[cfg(test)]
